requirements.txt
python-telegram-bot~=21.1.1
redis[hiredis]~=5.0.3
hiredis~=2.3.2
uvloop~=0.19.0
